refactor and comment form package
